Link compiler-rt in a default library search path
We used to install compiler-rt in the default location and then specify
the full path to it when we needed. This is annoying in certain
situations. We know create a symlink in a directory in the default
search path, specifically `usr/x86_64-gentoo-linux-musl/lib/` so that
compiler-rt can be linked simply using `-lclang_rt.builtins-x86_64`.